Extracted CameraPanelInspector from Components.js
