+++ src/js/shot-generator-new/Editor.js
@@ -14,12 +14,11 @@ const log = require('electron-log')
 const THREE = require('three')
 window.THREE = window.THREE || THREE
 require('../vendor/OutlineEffect')
-
+const CameraPanelInspector = require('./components/CameraPanelInspector').default
 const {
   SceneContext,
   ElementsPanel,
   CameraInspector,
-  CameraPanelInspector,
   BoardInspector,
   GuidesInspector,
   CamerasInspector,
